refactor - created separate route files
