+++ db/connections.js
@@ -1,0 +1,17 @@
+const mysql = require('mysql2');
+const mystuff = require('../utils/mystuff');
+
+// Connect to database
+const db = mysql.createConnection(
+    {
+      host: 'localhost',
+      // Your MySQL username,
+      user: 'root',
+      // Your MySQL password
+      password: mystuff,
+      database: 'election'
+    },
+    console.log('Connected to the election database.')
+);
+
+module.exports = db;
